core: extract engine error mapping
